requirements.txt
Django==4.1.7
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4